+++ doc/colors.md
@@ -1,0 +1,24 @@
+# COLORS
+
+## Default
+
+indigo lighten-2  
+indigo--text text--lighten-2
+
+indigo lighten-3  
+indigo--text text--lighten-3
+
+## Success
+
+green lighten-1  
+green--text text--lighten-1
+
+## Edition
+
+yellow darken-2  
+yellow--text text--darken-2
+
+## Suppression
+
+deep-orange lighten-1  
+deep-orange--text text--lighten-1
