+++ vuejs_app/src/views/Home.vue
@@ -10,3 +10,7 @@ export default {
   name: 'Home',
 }
 </script>
+
+<style lang="scss">
+
+</style>
